修复播放页收藏夹列表问题 (#1702)
* 修复播放页收藏夹列表问题 (#1666, #1669, #1694)

* Update FavoriteListResponse.cs

+++ src/Models/Models.BiliBili/Player/FavoriteListResponse.cs
@@ -34,19 +34,19 @@ namespace Bili.Models.BiliBili
         /// 收藏夹完整Id.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "id", Required = Required.Default)]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// 收藏夹原始Id.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "fid", Required = Required.Default)]
-        public int FolderId { get; set; }
+        public long FolderId { get; set; }
 
         /// <summary>
         /// 用户Id.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "mid", Required = Required.Default)]
-        public int UserId { get; set; }
+        public long UserId { get; set; }
 
         /// <summary>
         /// 收藏夹标题.
